erniesage: propeller & ernie

+++ examples/erniesage_propeller/preprocessing/dump_graph.py
@@ -25,34 +25,42 @@ import logging
 import multiprocessing
 from functools import partial
 from io import open
+import logging
 
 import numpy as np
 import tqdm
 import pgl
 from pgl.graph_kernel import alias_sample_build_table
-from pgl.utils.logger import log
+#from pgl.utils.logger import log
+import paddle.fluid.dygraph as D
+from easydict import EasyDict as edict
+import yaml
+from propeller import log
 
-from tokenization import FullTokenizer
+from ernie.tokenizing_ernie import ErnieTokenizer
+from ernie.tokenizing_ernie import ErnieTinyTokenizer
+from ernie.modeling_ernie import ErnieModel
+
+log.setLevel(logging.DEBUG)
 
 
 def term2id(string, tokenizer, max_seqlen):
-    #string = string.split("\t")[1]
     tokens = tokenizer.tokenize(string)
     ids = tokenizer.convert_tokens_to_ids(tokens)
     ids = ids[:max_seqlen-1]
-    ids = ids + [2] # ids + [sep]
-    ids = ids + [0] * (max_seqlen - len(ids))
+    ids = ids + [tokenizer.sep_id] # ids + [sep]
+    ids = ids + [tokenizer.pad_id] * (max_seqlen - len(ids))
     return ids
 
-def load_graph(args, str2id, term_file, terms, item_distribution):
+def load_graph(config, str2id, term_file, terms, item_distribution):
     edges = []
-    with io.open(args.graphpath, encoding=args.encoding) as f:
+    with io.open(config.graph_data, encoding=config.encoding) as f:
         for idx, line in enumerate(f):
             if idx % 100000 == 0:
-                log.info("%s readed %s lines" % (args.graphpath, idx))
+                log.info("%s readed %s lines" % (config.graph_data, idx))
             slots = []
             for col_idx, col in enumerate(line.strip("\n").split("\t")):
-                s = col[:args.max_seqlen]
+                s = col[:config.max_seqlen]
                 if s not in str2id:
                     str2id[s] = len(str2id)
                     term_file.write(str(col_idx) + "\t" + col + "\n")
@@ -67,16 +75,16 @@ def load_graph(args, str2id, term_file, terms, item_distribution):
     edges = np.array(edges, dtype="int64")
     return edges
 
-def load_link_predict_train_data(args, str2id, term_file, terms, item_distribution):
+def load_link_predict_train_data(config, str2id, term_file, terms, item_distribution):
     train_data = []
     neg_samples = []
-    with io.open(args.inpath, encoding=args.encoding) as f:
+    with io.open(config.train_data, encoding=config.encoding) as f:
         for idx, line in enumerate(f):
             if idx % 100000 == 0:
-                log.info("%s readed %s lines" % (args.inpath, idx))
+                log.info("%s readed %s lines" % (config.train_data, idx))
             slots = []
             for col_idx, col in enumerate(line.strip("\n").split("\t")):
-                s = col[:args.max_seqlen]
+                s = col[:config.max_seqlen]
                 if s not in str2id:
                     str2id[s] = len(str2id)
                     term_file.write(str(col_idx) + "\t" + col + "\n")
@@ -88,23 +96,23 @@ def load_link_predict_train_data(args, str2id, term_file, terms, item_distributi
             neg_samples.append(slots[2:])
             train_data.append((src, dst))
     train_data = np.array(train_data, dtype="int64")
-    np.save(os.path.join(args.outpath, "train_data.npy"), train_data)
+    np.save(os.path.join(config.graph_work_path, "train_data.npy"), train_data)
     if len(neg_samples) != 0:
-        np.save(os.path.join(args.outpath, "neg_samples.npy"), np.array(neg_samples))
+        np.save(os.path.join(config.graph_work_path, "neg_samples.npy"), np.array(neg_samples))
 
-def load_node_classification_train_data(args, str2id, term_file, terms, item_distribution):
+def load_node_classification_train_data(config, str2id, term_file, terms, item_distribution):
     train_data = []
     neg_samples = []
-    with io.open(args.inpath, encoding=args.encoding) as f:
+    with io.open(config.train_data, encoding=config.encoding) as f:
         for idx, line in enumerate(f):
             if idx % 100000 == 0:
-                log.info("%s readed %s lines" % (args.inpath, idx))
+                log.info("%s readed %s lines" % (config.train_data, idx))
             slots = []
             col_idx = 0
             slots = line.strip("\n").split("\t")
             col = slots[0]
             label = int(slots[1])
-            text = col[:args.max_seqlen]
+            text = col[:config.max_seqlen]
             if text not in str2id:
                 str2id[text] = len(str2id)
                 term_file.write(str(col_idx) + "\t" + col + "\n")
@@ -112,22 +120,22 @@ def load_node_classification_train_data(args, str2id, term_file, terms, item_dis
             src = str2id[text]
             train_data.append([src, label])
     train_data = np.array(train_data, dtype="int64")
-    np.save(os.path.join(args.outpath, "train_data.npy"), train_data)
+    np.save(os.path.join(config.graph_work_path, "train_data.npy"), train_data)
 
-def dump_graph(args):
-    if not os.path.exists(args.outpath):
-        os.makedirs(args.outpath)
+def dump_graph(config):
+    if not os.path.exists(config.graph_work_path):
+        os.makedirs(config.graph_work_path)
     str2id = dict()
-    term_file = io.open(os.path.join(args.outpath, "terms.txt"), "w", encoding=args.encoding)
+    term_file = io.open(os.path.join(config.graph_work_path, "terms.txt"), "w", encoding=config.encoding)
     terms = []
     item_distribution = []
 
-    edges = load_graph(args, str2id, term_file, terms, item_distribution)
-    #load_train_data(args, str2id, term_file, terms, item_distribution)
-    if args.task == "link_predict":
-        load_link_predict_train_data(args, str2id, term_file, terms, item_distribution)
-    elif args.task == "node_classification":
-        load_node_classification_train_data(args, str2id, term_file, terms, item_distribution)
+    edges = load_graph(config, str2id, term_file, terms, item_distribution)
+    #load_train_data(config, str2id, term_file, terms, item_distribution)
+    if config.task == "link_predict":
+        load_link_predict_train_data(config, str2id, term_file, terms, item_distribution)
+    elif config.task == "node_classification":
+        load_node_classification_train_data(config, str2id, term_file, terms, item_distribution)
     else:
         raise ValueError
 
@@ -140,36 +148,42 @@ def dump_graph(args):
     indegree = graph.indegree()
     graph.indegree()
     graph.outdegree()
-    graph.dump(args.outpath)
+    graph.dump(config.graph_work_path)
     
     # dump alias sample table
     item_distribution = np.array(item_distribution)
     item_distribution = np.sqrt(item_distribution)
     distribution = 1. * item_distribution / item_distribution.sum()
     alias, events = alias_sample_build_table(distribution)
-    np.save(os.path.join(args.outpath, "alias.npy"), alias)
-    np.save(os.path.join(args.outpath, "events.npy"), events)
+    np.save(os.path.join(config.graph_work_path, "alias.npy"), alias)
+    np.save(os.path.join(config.graph_work_path, "events.npy"), events)
     log.info("End Build Graph")
 
-def dump_node_feat(args):
+def dump_node_feat(config):
     log.info("Dump node feat starting...")
-    id2str = [line.strip("\n").split("\t")[-1] for line in io.open(os.path.join(args.outpath, "terms.txt"), encoding=args.encoding)]
-    pool = multiprocessing.Pool()
-    tokenizer = FullTokenizer(args.vocab_file)
-    term_ids = pool.map(partial(term2id, tokenizer=tokenizer, max_seqlen=args.max_seqlen), id2str)
-    np.save(os.path.join(args.outpath, "term_ids.npy"), np.array(term_ids, np.uint16))
+    id2str = [line.strip("\n").split("\t")[-1] for line in io.open(os.path.join(config.graph_work_path, "terms.txt"), encoding=config.encoding)]
+    if "tiny" in config.ernie_name:
+        tokenizer = ErnieTinyTokenizer.from_pretrained(config.ernie_name)
+        #tokenizer.vocab = tokenizer.sp_model.vocab
+        term_ids = [partial(term2id, tokenizer=tokenizer, max_seqlen=config.max_seqlen)(s) for s in id2str]
+    else:
+        tokenizer = ErnieTokenizer.from_pretrained(config.ernie_name)
+        pool = multiprocessing.Pool()
+        term_ids = pool.map(partial(term2id, tokenizer=tokenizer, max_seqlen=config.max_seqlen), id2str)
+        pool.terminate()
+    np.save(os.path.join(config.graph_work_path, "term_ids.npy"), np.array(term_ids, np.uint16))
     log.info("Dump node feat done.")
-    pool.terminate()
+
+def download_ernie_model(config):
+    D.guard().__enter__()
+    model = ErnieModel.from_pretrained(config.ernie_name)
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='main')
-    parser.add_argument("-i", "--inpath", type=str, default=None)
-    parser.add_argument("-g", "--graphpath", type=str, default=None)
-    parser.add_argument("-l", "--max_seqlen", type=int, default=30)
-    parser.add_argument("--vocab_file", type=str, default="./vocab.txt")
-    parser.add_argument("--encoding", type=str, default="utf8")
-    parser.add_argument("--task", type=str, default="link_predict", choices=["link_predict", "node_classification"])
-    parser.add_argument("-o", "--outpath", type=str, default=None)
+    parser.add_argument("--conf", type=str, default="./config.yaml")
     args = parser.parse_args()
-    dump_graph(args)
-    dump_node_feat(args)
+    config = edict(yaml.load(open(args.conf), Loader=yaml.FullLoader))
+
+    dump_graph(config)
+    dump_node_feat(config)
+    download_ernie_model(config)
